refactor(api-ui): prune redundant route handlers and deprecated wallet shims (Phase 5)
- Prune redundant router.patch moderation endpoints in backend/api/src/routes/adminRoutes.ts
- Decommission deprecated wallet shims in core/src/services/wallet/
- Migrate remaining callers to canonical payments domain exports
- Consolidate web ROUTES constants into apps/web/src/lib/routeUtils.ts and delete logic/routes.ts
- Ratchet down compatibility-marker-baseline from 126 to 122 tracked entries
