+++ apps/web/src/components/user/ListingDetail.tsx
@@ -22,7 +22,7 @@ import { AdPlacementSlot } from "@/components/common/AdPlacementSlot";
 
 import { canUserPerformAction } from "../../lib/logic/ownership";
 import { getActionBarVariant } from "../../lib/logic/bottomBarActions";
-import { ROUTES } from "../../lib/logic/routes";
+import { ROUTES } from "@/lib/routeUtils";
 import { useListingDetailQuery, useSavedAdsQuery } from "@/hooks/queries/useListingsQuery";
 import { useAuth } from "@/context/AuthContext";
 import { useViewTracking } from "./hooks/useViewTracking";
