+++ apps/web/src/components/user/listing-detail/AdBusinessCard.tsx
@@ -2,8 +2,7 @@ import { Badge } from "@esparex/ui";
 import { Button } from "@esparex/ui";
 import { Building2, Store, MapPin, Calendar, ExternalLink } from "@esparex/ui";
 import { notify } from "@/lib/feedback";
-import { ROUTES } from "@/lib/logic/routes";
-import type { AdDetailNavigateFn } from "@/lib/routeUtils";
+import { ROUTES, type AdDetailNavigateFn } from "@/lib/routeUtils";
 import type { Ad } from "@/schemas/ad.schema";
 import { formatStableDate } from "@/lib/formatters";
 import { resolveBusinessLocationLabel } from "@/lib/listings/listingPresentation";
