+++ scripts/policy/compatibility-marker-baseline.json
@@ -13,7 +13,6 @@
   "apps/web/src/lib/api/APIError.ts": 1,
   "apps/web/src/lib/api/result.ts": 1,
   "apps/web/src/lib/api/validateApiEnv.ts": 1,
-  "apps/web/src/lib/logic/routes.ts": 2,
   "apps/web/src/types/home.ts": 1,
   "apps/web/src/types/location.ts": 2,
   "backend/api/src/__tests__/config/validateEnv.spec.ts": 4,
@@ -57,7 +56,6 @@
   "backend/api/src/middleware/fraudMiddleware.ts": 1,
   "backend/api/src/middleware/requestId.ts": 2,
   "backend/api/src/middleware/validateIdOrSlug.ts": 4,
-  "backend/api/src/routes/adminRoutes.ts": 1,
   "backend/api/src/utils/apiResponse.ts": 1,
   "core/src/__tests__/services/AdminUsersService.spec.ts": 1,
   "core/src/__tests__/services/CatalogSearchGovernanceService.spec.ts": 2,
@@ -101,8 +99,6 @@
   "core/src/services/lifecycle/LifecycleGuard.ts": 1,
   "core/src/services/location/_shared/hierarchyLoader.ts": 3,
   "core/src/services/smartAlert/SmartAlertMutationService.ts": 1,
-  "core/src/services/wallet/WalletQueryService.ts": 1,
-  "core/src/services/wallet/WalletService.ts": 1,
   "core/src/utils/businessStatus.ts": 2,
   "core/src/utils/phoneUtils.ts": 1,
   "core/src/utils/s3.ts": 1,
